Add InputSource for new InputSystem
