+++ Assets/Code/src/Runtime/Input/TapPlayerInputSource.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dfbe42bc5f31acd4c84300c821837693
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
