+++ Assets/Code/src/Runtime/Input/UnityPlayerInputSource.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 405f32a53044a4d43925a3be653ec35d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
