+++ Assets/Resources/Config/GameplayConfig.asset
@@ -3,7 +3,7 @@
 --- !u!114 &11400000
 MonoBehaviour:
   m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
+  m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 0}
   m_Enabled: 1
@@ -11,6 +11,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 63c494f399c10ba4492cfe6e2b78abec, type: 3}
   m_Name: GameplayConfig
   m_EditorClassIdentifier: 
-  inputSource: HouraiTeahouse.FantasyCrescendo.InControlInputSource
+  inputSource: HouraiTeahouse.FantasyCrescendo.UnityInputSource
   _maxLedgeHangTime: 3
   _ledgeGrabCooldown: 3
