+++ ProjectSettings/ProjectSettings.asset
@@ -647,5 +647,5 @@ PlayerSettings:
   projectName: Fantasy Crescendo
   organizationId: james7132
   cloudEnabled: 0
-  enableNativePlatformBackendsForNewInputSystem: 0
+  enableNativePlatformBackendsForNewInputSystem: 1
   disableOldInputManagerSupport: 0
